Moved wdz load headers into include dir
